backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
confluent-kafka>=2.3.0
python-dotenv>=1.0.0
